[COMSUP-155] Move search to own component; fix query hash for non-guides (#909)
* Move search to own component; fix query hash for non-guides

* Update callout for `set_done`
